move to clip.py
